# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ghost Protocol API",
    "version": "1.0.0",
    "description": "Machine-readable API surface for GhostRank discovery and GhostGate access."
  },
  "servers": [
    {
      "url": "https://ghostprotocol.cc"
    }
  ],
  "paths": {
...
    "/api/gate/{service}": {
      "get": {
        "summary": "Authorize access and consume credits",
        "operationId": "gateAuthorizeGet",
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-ghost-sig",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-ghost-payload",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-ghost-credit-cost",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized"
          },
          "400": {
            "description": "Malformed auth"
          },
          "401": {
            "description": "Invalid signature / service mismatch / expired"
          },
          "402": {
            "description": "Insufficient credits"
          },
          "409": {
            "description": "Replay detected"
          },
          "423": {
            "description": "Service not live"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "post": {
        "summary": "Authorize access and consume credits",
        "operationId": "gateAuthorizePost",
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-ghost-sig",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-ghost-payload",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-ghost-credit-cost",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized"
          },
          "400": {
            "description": "Malformed auth"
          },
          "401": {
            "description": "Invalid signature / service mismatch / expired"
          },
          "402": {
            "description": "Insufficient credits"
          },
          "409": {
            "description": "Replay detected"
          },
          "423": {
            "description": "Service not live"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
  }
}